frustration breaking point commit - basic functionality in place - still no working arrays or timestamps
Things I have learned today:

C is absolutely horrible. No, it is.
Working with strings in C is a joke
Getting a timestamp and using an array in C is a joke
Why anyone would choose C for a platform is beyond me.
Everything is significantly more complex than it needs to be.