Major update today
Finally had time to get back to this (and desire because I was pretty
over it).

I have branched to two code-bases - one for what will become the remote
solution in the long run when I get help with the C code and one which
works more simply and makes a dangerous assumption of “always or at
least mostly always” connected.

Thanks to my wife for suggesting I not fix a problem I don’t have yet…